refactor: remove skew card component files and update SkewCardData to load setup and usage from markdown files
