+++ lib/components/presentation/animations/skew_card/usage.md
@@ -1,0 +1,15 @@
+# Usage Example
+
+```dart
+SkewCard(
+  title: 'Hello',
+  image: AssetImage('assets/skew_card/sample.png'),
+  onTap: () => print('Card tapped!'),
+)
+```
+
+## Best Practices
+
+- Use high-resolution images for best appearance.
+- Wrap in a `Padding` widget for spacing.
+- Use meaningful titles for accessibility.
